Changed the variable type from double to decimal

+++ 02.DataTypesAndVariables/ExactSumOfNumbers/Program.cs
@@ -7,11 +7,11 @@ namespace ExactSumOfNumbers
         static void Main(string[] args)
         {
             int n = int.Parse(Console.ReadLine());
-            double sum = 0;
+            decimal sum = 0;
 
             for (int i = 0; i < n; i++)
             {
-                double number = double.Parse(Console.ReadLine());
+                decimal number = decimal.Parse(Console.ReadLine());
                 sum += number;
             }
